refactor(g2g-transfer): co-locate the install-gate middleware with its only consumer
`generateAdminRequiredIfInstalled` encodes a g2g-transfer-specific rule ("admin
required unless the app is not yet installed", needed only so the transfer
receiver can accept a transfer into a fresh, not-yet-installed instance). It has
no other consumer and no realistic shared use, so the general middlewares/
directory overstated its generality. Move it next to the route as a flat sibling,
symmetric with g2g-transfer-attachment-metadata.ts. No behavior change.
