+++ apps/app/src/server/routes/apiv3/g2g-transfer-admin-required-if-installed.ts
